调研问题返回qid 0 bug修改

+++ s2si/src/com/joyque/dao/ISurveyQuestionDao.java
@@ -10,7 +10,7 @@ public interface ISurveyQuestionDao {
 	
 	public SurveyQuestion GetSurveyQuestion(long qid);
 	
-	public int insertSurveyQuestion(SurveyQuestion info);
+	public long insertSurveyQuestion(SurveyQuestion info);
 	
 	public int updateSurveyQuestion(SurveyQuestion info);
